Move hotkeys too `App.vue`
